requirements.txt
telethon>=1.24.0
python-dotenv>=0.19.0
pandas>=1.3.0
requests>=2.25.0
python-dateutil>=2.8.0